Update .config/discord/settings.json Update gnome_monitors.zsh

+++ run_after_gnome_monitors.zsh
@@ -5,6 +5,8 @@ set -o errexit -o nounset
 from=$(realpath ~/.config/monitors.xml)
 to="/var/lib/gdm/.config/monitors.xml"
 
+[ ! -f "$from" ] && exit 0
+
 cmp --silent $from $to || (
     echo "$from changed, copying to $to..."
     sudo cp $from $to
